refactored code and uncommented thread spawn
