+++ closures/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod shirtStock;
+pub mod shirt_stock;
 pub mod closure_ops;
